return space parent/children as actual child nodes, as defined in the graphql query
