+++ src/graphql/schema.gql
@@ -222,8 +222,8 @@ type Space {
   validation: Strategy
   followersCount: Int
   proposalsCount: Int
-  parent: String
-  children: [String]
+  parent: Space
+  children: [Space]
 }
 
 type SpaceFilters {
